samples: radio_test: Indicate link status in cont. Rx
When running a continuous RX test (without number of packets
specified), indicate with LED0 lit when the packets are being
received with correct CRC. Turn off the LED when PER > 10% or
no correct packet is received in default timeout (100ms).

Ref. NCSDK-36346

temporary solution for radio_test to indicate the link status
in the continuous RX mode.

Signed-off-by: Michał Grochala <michal.grochala@nordicsemi.no>